fix clang tidy cast warning

+++ src/core/qgstessellator.cpp
@@ -584,7 +584,7 @@ void QgsTessellator::addTriangleVertices(
 
     // cppcheck-suppress negativeContainerIndex
     QVector3D point = points[ index ];
-    const float z = mNoZ ? 0.0 : point.z();
+    const double z = mNoZ ? 0.0 : point.z();
     QVector4D pt( point.x(), point.y(), z, 0 );
 
     pt = *transformMatrix * pt;
